Support firmware update
With this commit we expose neccessary operations for UI to be
able to render a functional "Firmware Update" modal. User has
to pick a firmware binary from the drop-down to apply it to
selected server(s). Eventually, when she submits the modal, a
new Request of type PhysicalServerFirmwareUpdateRequest is created
by means of POSTing to `/api/request`.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>